refactor: Rename service activation.ts to strategy.ts
This changes the name to be more specific
with the current architecture
