+++ lsp/src/service/filetypes/buildspec/strategy.ts
